# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice-Based Bus Transportation System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  USER: "USER",
  ADMIN: "ADMIN",
} as const;

export interface RegisterRequest {
  username: string;
  password: string;
  role?: RegisterRequestRole;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  USER: "USER",
  ADMIN: "ADMIN",
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
}

export interface LoginResponse {
  user: User;
  token: string;
}

export interface Bus {
  id: number;
  busNumber: string;
  fromLocation: string;
  toLocation: string;
  departureTime: string;
  arrivalTime: string;
  totalSeats: number;
  availableSeats: number;
  price: number;
}

export interface SeatInfo {
  seatNumber: number;
  isBooked: boolean;
}

export type BusDetail = Bus & {
  seats: SeatInfo[];
};

export interface CreateBusRequest {
  busNumber: string;
  fromLocation: string;
  toLocation: string;
  departureTime: string;
  arrivalTime: string;
  totalSeats: number;
  price: number;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  CONFIRMED: "CONFIRMED",
  CANCELLED: "CANCELLED",
} as const;

export interface Booking {
  id: number;
  userId: number;
  busId: number;
  seatNumbers: string;
  bookingTime: string;
  status: BookingStatus;
  bus?: Bus;
  username?: string;
}

export interface CreateBookingRequest {
  userId: number;
  busId: number;
  seatNumbers: string;
}

export interface StatsOverview {
  totalBuses: number;
  totalBookings: number;
  totalUsers: number;
  activeRoutes: number;
}

export interface PopularRoute {
  fromLocation: string;
  toLocation: string;
  bookingCount: number;
  busCount: number;
}

export type SearchBusesParams = {
  /**
   * Departure location
   */
  from?: string;
  /**
   * Arrival location
   */
  to?: string;
  /**
   * Bus number to search for
   */
  busNumber?: string;
  /**
   * Free-text or voice query
   */
  query?: string;
};

export type ListBookingsParams = {
  /**
   * Filter by user ID
   */
  userId?: number;
};