# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice-Based Bus Transportation System API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";
import type { Bus } from "./bus";

export interface Booking {
  id: number;
  userId: number;
  busId: number;
  seatNumbers: string;
  bookingTime: string;
  status: BookingStatus;
  bus?: Bus;
  username?: string;
}