# lib/api-zod/src/generated/types/busDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice-Based Bus Transportation System API
 * OpenAPI spec version: 0.1.0
 */
import type { Bus } from "./bus";
import type { SeatInfo } from "./seatInfo";

export type BusDetail = Bus & {
  seats: SeatInfo[];
};