# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice-Based Bus Transportation System API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterRequestRole } from "./registerRequestRole";

export interface RegisterRequest {
  username: string;
  password: string;
  role?: RegisterRequestRole;
}