# lib/api-zod/src/generated/types/seatInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice-Based Bus Transportation System API
 * OpenAPI spec version: 0.1.0
 */

export interface SeatInfo {
  seatNumber: number;
  isBooked: boolean;
}