# lib/api-zod/src/generated/types/statsOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice-Based Bus Transportation System API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsOverview {
  totalBuses: number;
  totalBookings: number;
  totalUsers: number;
  activeRoutes: number;
}